# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    TaskStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, REVIEW_REQUIRED, COMPLETED, NEEDS_FIX, BLOCKED, CANCELLED]
    VerificationVerdict:
      type: string
      enum: [APPROVED, NOT_APPROVED, BLOCKED]
...
    Task:
      type: object
      required: [id, title, objective, status, dependencies, workspaceId, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        objective:
          type: string
        scope:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        verificationCommands:
          type: array
          items:
            type: string
        assignedTo:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        dependencies:
          type: array
          items:
            type: string
        parallelGroup:
          type: string
        workspaceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completionSummary:
          type: string
        verificationVerdict:
          $ref: "#/components/schemas/VerificationVerdict"
        verificationReport:
          type: string
...
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
          const: true

# ─────────────────────────────────────────────────────────
# API Paths
# ─────────────────────────────────────────────────────────
paths:
...
  /api/tasks:
    get:
      operationId: listTasks
      summary: List tasks
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: assignedTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, objective]
              properties:
                title:
                  type: string
                objective:
                  type: string
                workspaceId:
                  type: string
                  default: "default"
                scope:
                  type: string
                acceptanceCriteria:
                  type: array
                  items:
                    type: string
                verificationCommands:
                  type: array
                  items:
                    type: string
                dependencies:
                  type: array
                  items:
                    type: string
                parallelGroup:
                  type: string
      responses:
        "200":
          description: Created task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteAllTasks
      summary: Delete all tasks for a workspace
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
  /api/tasks/{id}:
    get:
      operationId: getTask
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
  /api/tasks/{id}/status:
    post:
      operationId: updateTaskStatus
      summary: Update task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/TaskStatus"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
...
  /api/a2a/rpc:
    post:
      operationId: a2aRpc
      summary: A2A JSON-RPC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: JSON-RPC response
    get:
      operationId: a2aRpcSse
      summary: A2A SSE stream
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}

  # ── GitHub Virtual Workspace ──
...
  /api/rpc:
    post:
      operationId: rpcCall
      summary: Generic JSON-RPC endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method:
                  type: string
                params:
                  type: object
      responses:
        "200":
          description: RPC result

  # ── Traces ──
...
  /api/background-tasks:
    get:
      operationId: listBackgroundTasks
      summary: List background tasks
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Background task list
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: createBackgroundTask
      summary: Create a background task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Created task
...
  /api/background-tasks/{id}:
    get:
      operationId: getBackgroundTask
      summary: Get a background task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task detail
    delete:
      operationId: cancelBackgroundTask
      summary: Cancel a background task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
  /api/background-tasks/{id}/retry:
    post:
      operationId: retryBackgroundTask
      summary: Retry a failed background task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retry queued

  # ── Schedules sub-operations ──
...
  /api/a2a/tasks/{id}:
    get:
      operationId: getA2ATask
      summary: Get an A2A task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task detail
    post:
      operationId: updateA2ATask
      summary: Update / respond to an A2A task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated

  # ── RPC discovery (Rust) ──
...